4.2

+++ ivp/src/app_alogscan/ScanHandler.cpp
@@ -1,16 +1,27 @@
 /*****************************************************************/
-/*    NAME: Michael Benjamin                                     */
-/*    ORGN: NAVSEA Newport RI and MIT Cambridge MA               */
+/*    NAME: Michael Benjamin, Henrik Schmidt, and John Leonard   */
+/*    ORGN: Dept of Mechanical Eng / CSAIL, MIT Cambridge MA     */
 /*    FILE: ScanHandle.cpp                                       */
 /*    DATE: June 5rd, 2008                                       */
 /*                                                               */
-/* This is unreleased BETA code. No permission is granted or     */
-/* implied to use, copy, modify, and distribute this software    */
-/* except by the author(s).                                      */
+/* This program is free software; you can redistribute it and/or */
+/* modify it under the terms of the GNU General Public License   */
+/* as published by the Free Software Foundation; either version  */
+/* 2 of the License, or (at your option) any later version.      */
+/*                                                               */
+/* This program is distributed in the hope that it will be       */
+/* useful, but WITHOUT ANY WARRANTY; without even the implied    */
+/* warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR       */
+/* PURPOSE. See the GNU General Public License for more details. */
+/*                                                               */
+/* You should have received a copy of the GNU General Public     */
+/* License along with this program; if not, write to the Free    */
+/* Software Foundation, Inc., 59 Temple Place - Suite 330,       */
+/* Boston, MA 02111-1307, USA.                                   */
 /*****************************************************************/
 
 #include <iostream>
-#include <math.h>
+#include <cmath>
 #include <cstdio>
 #include "MBUtils.h"
 #include "ALogScanner.h"
@@ -108,8 +119,8 @@ void ScanHandler::handle(const string& alogfile)
   int start_digits = (int)(log10( m_report.getMaxStartTime())) + 4;
   int stop_digits  = (int)(log10( m_report.getMaxStopTime())) + 4;
 #else 
-  int line_digits  = (int)(log10(m_report.getMaxLines())) + 2;
-  int char_digits  = (int)(log10(m_report.getMaxChars())) + 2;
+  int line_digits  = static_cast<int>(log10(static_cast<double>(m_report.getMaxLines()))) + 2;
+  int char_digits  = static_cast<int>(log10(static_cast<double>(m_report.getMaxChars()))) + 2;
   int start_digits = (int)(log10(m_report.getMaxStartTime())) + 4;
   int stop_digits  = (int)(log10(m_report.getMaxStopTime())) + 4;
 #endif 
@@ -250,3 +261,4 @@ void ScanHandler::appStatReport()
 	   s_chars.c_str(), s_lines_pct.c_str(), s_chars_pct.c_str());
   }  
 }
+
